introduced post html interaction handler to avoid and actions duplications
