+++ src/components/postHtmlRenderer/index.ts
@@ -1,1 +1,2 @@
 export * from './postHtmlRenderer';
+export * from './postInteractionHandler';
